+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectOverviewView.groovy
@@ -15,6 +15,7 @@ import com.vaadin.ui.themes.ValoTheme
 import groovy.util.logging.Log4j2
 import life.qbic.portal.sampletracking.Constants
 import life.qbic.portal.sampletracking.communication.notification.NotificationService
+import life.qbic.portal.sampletracking.components.GridUtils
 import life.qbic.portal.sampletracking.components.projectoverview.download.DownloadProjectController
 import life.qbic.portal.sampletracking.components.projectoverview.samplelist.FailedQCSamplesView
 import life.qbic.portal.sampletracking.components.projectoverview.samplelist.ProjectOverviewController
@@ -43,6 +44,7 @@ class ProjectOverviewView extends VerticalLayout {
 
     private Grid<ProjectSummary> projectGrid
     private HorizontalSplitPanel splitPanel
+    private static final Collection<String> columnIdsWithFilters = ["ProjectCode", "ProjectTitle"]
 
     final static int MAX_CODE_COLUMN_WIDTH = 400
     final static int MAX_STATUS_COLUMN_WIDTH = 200
@@ -299,8 +301,8 @@ class ProjectOverviewView extends VerticalLayout {
         if( viewModel.selectedProject ) {
             projectGrid.select(viewModel.selectedProject)
         }
+        GridUtils.setupFilters(projectGrid, columnIdsWithFilters)
     }
-
 
     private void tryToDownloadManifest() {
         Optional<ProjectSummary> selectedSummary = Optional.empty()
@@ -383,19 +385,15 @@ class ProjectOverviewView extends VerticalLayout {
      * @param sampleCount The total number of samples registered
      */
     private static State determineCompleteness(SampleCount sampleCount) {
-        if (sampleCount.failingSamples > 0){
+        if (sampleCount.failingSamples > 0) {
             return State.FAILED
-        }
-        else if (sampleCount.passingSamples == sampleCount.totalSampleCount) {
+        } else if (sampleCount.passingSamples == sampleCount.totalSampleCount) {
             return State.COMPLETED
-        }
-        else if (sampleCount.passingSamples < sampleCount.totalSampleCount) {
+        } else if (sampleCount.passingSamples < sampleCount.totalSampleCount) {
             return State.IN_PROGRESS
-        }
-        else {
+        } else {
             //unexpected!!
             throw new IllegalStateException("status count $sampleCount.passingSamples must not be greater total count $sampleCount.totalSampleCount")
         }
     }
-
 }